Added persistent login with jsonwebtoken

+++ server.js
@@ -5,6 +5,7 @@ const cors = require("cors")
 const app = express()
 const db = require("./db")
 const bcrypt = require('bcrypt')
+const jwt = require("jsonwebtoken")
 
 app.use(cors())
 app.use(express.json())
@@ -12,12 +13,6 @@ app.use(express.json())
 app.get('/', (req, res) => {
     res.send('Homepage here')
 })
-
-// TODO: Implement login system
-
-// TODO: Improve layout with bootstrap and other CSS rules
-
-// TODO: Complete README
 
 // Blog routes
 app.get('/api/blog', async (req, res) => {
@@ -54,8 +49,12 @@ app.get('/api/blog/peek/:count', async (req, res) => {
     }
 })
 
-app.post('/api/blog', async (req, res) => {
-    //Create new blog entry here
+//Create new blog entry
+app.post('/api/blog', verifyToken, async (req, res) => {
+    if(req.valid == false) {
+        return res.status(401);
+    }
+
     try {
         const result = await db.query("INSERT INTO blog (title, category, content) values ($1, $2, $3) returning *", 
         [req.body.title, req.body.category, req.body.content])
@@ -72,8 +71,12 @@ app.post('/api/blog', async (req, res) => {
     }
 })
 
-app.put('/api/blog/:id', async (req, res) => {
-    //Modify blog entry here
+//Modify blog entry
+app.put('/api/blog/:id', verifyToken, async (req, res) => {
+    if(req.valid == false) {
+        return res.status(401)
+    }
+    
     try {
         const result = await db.query("UPDATE blog SET title=$1, category=$2, content=$3 WHERE id=$4 returning *", 
         [req.body.title, req.body.category, req.body.content, req.params.id])
@@ -90,8 +93,8 @@ app.put('/api/blog/:id', async (req, res) => {
     }
 })
 
+// Show full blog entry
 app.get('/api/blog/:id', async (req, res) => {
-    //Show blog content here
     try {
         const result = await db.query("SELECT * FROM blog WHERE id=$1", [req.params.id])
 
@@ -111,7 +114,12 @@ app.get('/api/blog/:id', async (req, res) => {
     }
 })
 
-app.delete('/api/blog/:id', async (req, res) => {
+// Delete blog entry
+app.delete('/api/blog/:id', verifyToken, async (req, res) => {
+    if(req.valid == false) {
+        return res.status(401);
+    }
+    
     try {
         const result = await db.query("DELETE FROM blog WHERE id=$1", [req.params.id])
 
@@ -136,22 +144,26 @@ app.get('/api/projects', async (req, res) => {
 // Create user if no user yet
 app.post('/api/users', async (req, res) => {
     // 1. Check if a user exists in DB already
-    const userCount = await db.query('SELECT COUNT(*) FROM users')
+    try {
+        const userCount = await db.query('SELECT COUNT(*) FROM users')
 
-    // 2. If not, then create user based on contents of .env
-    let username = process.env.CUSER;
-    let password = process.env.CPASSWORD;
+        // 2. If not, then create user based on contents of .env
+        let username = process.env.CUSER;
+        let password = process.env.CPASSWORD;
 
-    // Hash password using bcrypt
-    const hashedPassword = await bcrypt.hash(password, 10)
+        // Hash password using bcrypt
+        const hashedPassword = await bcrypt.hash(password, 10)
 
-    // 3. Store that user in DB
-    if(userCount.rows[0].count == 0) {
-        await db.query("INSERT INTO users (Username, Password) values ($1, $2) returning *", 
-        [username, hashedPassword])
+        // 3. Store that user in DB
+        if(userCount.rows[0].count == 0) {
+            await db.query("INSERT INTO users (Username, Password) values ($1, $2) returning *", 
+            [username, hashedPassword])
+        }
+
+        res.status(200);
+    } catch (err) {
+        res.status(400);
     }
-
-    res.status(200);
 })
 
 // Login route
@@ -164,27 +176,51 @@ app.post('/api/login', async (req, res) => {
         const pass = await db.query('SELECT password FROM users WHERE username=$1', 
         [req.body.username])
 
-        if (pass.rows[0].password == null) {
-            return res.status(400).send('Cannot find user')
+        if (pass.rowCount == 0) {
+            return res.status(200).json({ accessToken: 'invalid_token' })
         }
-
-        console.log("username: success")
 
         if(await bcrypt.compare(req.body.password, pass.rows[0].password)) {
             validCredentials = true;
-            console.log("password: success")
+        } else {
+            return res.status(200).json({ accessToken: 'invalid_token' })
         }
 
         // 3. Generate a JWT (see video on jsonwebtoken)
         if (validCredentials) {
-
+            const user = { name: req.body.username }
+            const accessToken = jwt.sign(user, process.env.ACCESS_TOKEN_SECRET)
+            res.status(200).json({ accessToken })
         }
+
     } catch (err) {
         res.status(500).send();
     }
-
-    res.status(200).json();
 })
+
+// Authenticates a Json Web Token //TODO: Make verificationexpire?
+function verifyToken(req, res, next) {
+    const authHeader = req.headers['authorization']
+    const token = authHeader && authHeader.split(' ')[1]
+    req.valid = false;
+
+    if(token == null) {
+        console.log("token was NULL")
+        return res.sendStatus(401);
+    }
+
+    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
+        if(err) {
+            console.log("authentication FAILED")
+            return res.sendStatus(403);
+        } else {
+            console.log("authentication PASSED")
+            req.valid = true;
+        }
+    })
+
+    next();
+}
 
 const port = process.env.PORT
 app.listen(port, () => {
